t9114-git-svn-dcommit-merge.sh: use the $( ... ) construct for command substitution
The Git CodingGuidelines prefer the $(...) construct for command
substitution instead of using the backquotes `...`.

The backquoted form is the traditional method for command
substitution, and is supported by POSIX.  However, all but the
simplest uses become complicated quickly.  In particular, embedded
command substitutions and/or the use of double quotes require
careful escaping with the backslash character.

The patch was generated by:

   for _f in $(find . -name "*.sh")
   do
       sed -i 's@`\(.*\)`@$(\1)@g' ${_f}
   done

and then carefully proof-read.

Signed-off-by: Elia Pinto <gitter.spiros@gmail.com>